refactor(domain): Person 클래스 추가
이름 String을 포장하기 위해 클래스 분리 및 테스트 분리
