+++ src/test/java/ladder/domain/PeopleTest.java
@@ -8,8 +8,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 class PeopleTest {
 
@@ -26,14 +24,6 @@ class PeopleTest {
         assertThatThrownBy(() -> new People(
                 List.of("명오", "명오")
         ))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("이름이 1~5글자 범위를 벗어나면 예외를 발생한다.")
-    @ValueSource(strings = {"", "우아한테크코스"})
-    @ParameterizedTest
-    void nameLengthTest(String name) {
-        assertThatThrownBy(() -> new People(List.of(name)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
